backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.3.0
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.0
seaborn==0.13.2
joblib==1.4.2
pyyaml==6.0.1
httpx==0.27.0
pytest==8.2.1
pytest-asyncio==0.23.7
